Conditional Independency test only checks the left/right members.
Conditional indpendencies are only canonical to the left/right node.  Many different combinations
of 'givens' are valid as long as the set of givens makes it so left and right are d-separated.
Since the sources of the tests don't actually specify how the chose between differents sets of givens,
this test gets closer to the essence of the problem