Only returns FIRST left/right/conditions set from the d-separation iterator by default.
Since the powerset iterator first returns all sets of size one, then all sets of size 2, etc. this will find the
one of the simplest d-separation condition sets for a given left/right pair. In most cases, that is all that is
required.

With this optimization in place, tests run in half the time (3.7s vs. 8.6s).  HOWEVER, this pre-filter may cause
issues if an exhaustive exploration is desirable (for example, if a topological condition on the d-separation needs
to be met in a later step).  This chance of missing important cases is why it is controlled by a flag.