Automate checking binary/continous data (#194)
As a follow-up to #193, this PR extends the default conditional
independency test from just being the Cressie-Read test to dynamically
checking if the data is binary or not. If it's binary, it stays with
cressie read but if it's continuous, it sets it to the Pearson test.

This implementation doesn't yet consider discrete data with more than 2
values. The idea Pruthvi had in `eliater` was to manually specify a
threshold. I don't like this since it's parametric, and if you already
know what's going on in your data, you might as well set the test
correctly. An idea for an automated test is in here for future
reference:
https://stats.stackexchange.com/questions/12273/how-to-test-if-my-data-is-discrete-or-continuous