Update display of signs on variables in dsl (#220)
@djinnome @callahanr this PR fixes some issues with the DSL in the way
that it displays the status of variables. Most importantly, it now
differentiates between star is None, star is True, and star is False for
regular variables. This makes a pretty big simplification for
counterfactual transport's inputs since you can directly say things like

```python
-Infected @ +Vaccinated
```

instead of needing to have a 2-tuple

```python
(Infected @ +Vaccinated, -Infected)
```

Note: this was already possible, but the DSL didn't show it correctly

This PR also fixes a few bugs in the way population probabilities are
displayed